move SurveyorSelect to own file to prevent re-render on every select change
